==> 01_Lyon_DataPrep_Solution.ipynb <==
# Predicting the occupancies of Belgian trains

## 0. Tips and tricks Jupyter notebooks

* Export your notebooks locally: File -> Download as -> html/ipynb (do both!) (NOTE: don't trust the virtual environment, always download your latest version locally!)

* Shift + Enter to run a cell (or play button)

* TAB for code completion!

* Shells can be Code but also Markdown! (dropdown menu on top allows you to choose)

* Cell -> Run All might come in useful

* In case of problems you can always Kernel -> Restart

* In the Jupyter start window you have a 'new' button on the upper right => to open a terminal, a new python3 notebook, etc.

* You cannot upload very large files (>3GB) via the upload in Jupyter (bug), the datasets will always be available via een public Amazon url, so use wget or scp to get your data in /mnt on the virtual wall!


## 1. Data Preparation

### Essential Libraries for Python Data Science

#vector/matrix library
import numpy as np
#data frame library (similar to R)
import pandas as pd

#visualization library
%matplotlib inline
import matplotlib.pyplot as plt
plt.style.use('ggplot')

#regular expression library for data cleasning
import re

### Detour: Get up to speed with python / pandas

* If you are new to python have a look at the practicelab.ipynb in the github repo
* If you are new to pandas take some time for:
    - <a href="http://pandas.pydata.org/pandas-docs/stable/10min.html"> Pandas basics in 10 minutes </a>
    - <a href="http://pandas.pydata.org/pandas-docs/stable/visualization.html"> Pandas for data visualizations </a>
    - Use the pandas cheat sheet in the github repo
    
    
#### TIP: Pandas dataframes are immutable, so appending a row creates a copies the old df in the new one => creating a df like this is an O(N^2) operation (I learnt this the hard way)    
 


### 1a. Dataset characteristics

1. Read the train and test datasets
2. How many records and how many features are in train and test sets?
3. What is the range of the querytime column (earliest + latest date)
4. Merge training and test data (data cleansing will be the same for both)
5. Drop columns querytype and user_agent since they contain no useful info, this will speedup future df calculations

#1
path_train = './data/training_data.nldjson'
path_test  = './data/test.nldjson' 

#warning the input files are slightly different format, training data is in new-line delimited json, 
#testdata in regular json
df_train = pd.read_json(path_train, lines=True)
df_test = pd.read_json(path_test)

#to have a look at a dataframe just use head or tail
df_train.head(n=5)

df_test.head(n=5)

#2
print("Number of records in training set: " + str(df_train.shape[0]))
print("Number of records in test set: " + str(df_test.shape[0]))



print("Features training data: " + str(list(df_train.columns)))
print("Features test data: " + str(list(df_test.columns)))

#3
print('Querytime min (training): ' + str(df_train['querytime'].min()))
print('Querytime max (training): ' + str(df_train['querytime'].max()))
print()
print('Querytime min (test): ' + str(df_test['querytime'].min()))
print('Querytime max (test): ' + str(df_test['querytime'].max()))

#4
#merge train and test data AND reset the index
dataset_v1 = pd.concat([df_train, df_test]).reset_index(drop=True)    
dataset_v1.head(n=5)

#5
dataset_v2 = dataset_v1.drop(['querytype','user_agent'], axis=1)
dataset_v2.head(n=5)

### 1b. Parsing the columns

1. In the id column replace the NaNs with -1
2. Select a row by id and have a closer look at the post column (json object)
    * check multiple rows since the post object doesn't always have the same fields
3. Iterate over the frame while keeping track of the possible fields in post => what is the set of ps
4. Write a function to extract a field of the json object 
5. For every field in the json objects create an additional column, finally drop the post column
6. The query time column is now interpreted as a string => convert to datetime, the id column can be cast to integer



    


#1
transformed_column = dataset_v2['id'].apply(lambda i: -1 if pd.isnull(i) else i)
dataset_v2['id'] = transformed_column
dataset_v2.head(n=3)

#2
row = dataset_v2.iloc[2]
post_obj = row['post']
post_obj

#3
all_keys = []
for r in dataset_v2.iterrows():
    new_keys = list(r[1]['post'].keys())
    all_keys.extend(new_keys)
    
unique_keys = set(all_keys)
unique_keys

#4
def extract_field(json, field):
    if field in json:
        return json[field]
    else:
        return None

#test
dataset_v2['post'].apply(lambda j: extract_field(j, 'connection')).head()

#5
for k in unique_keys:
    new_col = dataset_v2['post'].apply(lambda j: extract_field(j, k))
    dataset_v2[k] = new_col
    
dataset_v3 = dataset_v2.drop('post', axis=1)    
dataset_v3.head()    

#6
dataset_v3['querytime'] = pd.to_datetime(dataset_v3['querytime'])

#Later onderscheid tussen winter en zomeruur vermijden door niet naar UTC te gaan
#dataset_v3['querytime' = pd.to_datetime(dataset_v3['querytime'].apply(lambda t: t.split("+")[0]))

dataset_v3['id'] = pd.to_numeric(dataset_v3['id'], downcast='integer')


dataset_v3.dtypes

### 1.3 And some more cleansing

- TASK: go over each of the columns, further parse them and pay close attention to NULL values
- HINT: don't mindlessly drop rows with missing data, the dataset has some level of redundancy so pay close attention!


1. How many NULL values per column?
2. Rework the vehicle column:
    - use the following information on the slides and at: https://nl.wikipedia.org/wiki/Lijst_van_treincategorie%C3%ABn_in_Belgi%C3%AB
    - extract connection type, train series, sequence number and you can also infer the direction of the train
    - keep checking for NULLs, some can be disguised (HINT: use the value_counts() function to check the distribution of values in a column)

4. connection
5. date
6. occupancy
7. to, from columns

#1

pd.isnull(dataset_v3['vehicle'])

for col in dataset_v3.columns:
    print(col + "\t"+ str(dataset_v3[pd.isnull(dataset_v3[col])].shape[0]))

dataset_v3[pd.isnull(dataset_v3['to'])]

#2
def remove_url_prefix(s):
    return s.split('/')[-1]

dataset_v3['vehicle'] = dataset_v3['vehicle'].apply(remove_url_prefix)
dataset_v3['vehicle'].value_counts()

def remove_irregulars(s):
    if pd.isnull(s):
        return s
    else:
        if s=='(null)':
            return None
        else:
            return s

dataset_v3['vehicle'] = dataset_v3['vehicle'].apply(remove_irregulars)
dataset_v3['vehicle'].value_counts()
dataset_v3[pd.isnull(dataset_v3['vehicle'])].head(n=5)

#INZICHT vehicle types zitten ook in connection => daar extraheren

dataset_v3.iloc[0]['connection']

def extract_train_series_id(r):
    series_number = int(re.sub(r"[a-zA-Z]", "", r)) 
    return series_number % 100 

def extract_train_series(r):
    series_number = int(re.sub(r"[a-zA-Z]", "", r)) 
    return series_number // 100 * 100
               
def extract_train_direction(r):
    series_id = extract_train_series_id(r)
    series_div = (series_id // 25) % 2
    return 'away' if series_div == 0 else 'back'



def extract_train_type(r):
    return re.sub(r"[0-9]", "", r.upper())

#3

dataset_v4 = dataset_v3.drop(['vehicle','date'],axis=1)
dataset_v4['vehicle'] = dataset_v4['connection'].apply(lambda r: r.split('/')[-1])
dataset_v4['vehicle'].value_counts()

dataset_v4['train_series']    = dataset_v4['vehicle'].apply(extract_train_series)
dataset_v4['train_direction'] = dataset_v4['vehicle'].apply(extract_train_direction)
dataset_v4['train_type']      = dataset_v4['vehicle'].apply(extract_train_type)


dataset_v5 = dataset_v4.drop(['vehicle','connection'], axis=1)

#4 date can just be dropped, querytime is much easier to work with

#5
dataset_v5['occupancy'] = dataset_v5['occupancy'].apply(lambda r: r.split('/')[-1] if pd.notnull(r) else None)
dataset_v5

#6
dataset_v5['from'].describe()

dataset_v5['to'].describe()

mask = pd.isnull(dataset_v5['to'])
dataset_v5[mask]

dataset_v6 = dataset_v5.drop(dataset_v5[mask].index)

dataset_v6['from'].describe()


dataset_v6['from'] = dataset_v6['from'].apply(lambda r: r.strip().split('/')[-1])
dataset_v6['to'] = dataset_v6['to'].apply(lambda r: r.strip().split('/')[-1])


mask = dataset_v6['from'].str.len() != 9
print(dataset_v6[mask].shape)
dataset_v6[mask]


dataset_v7 = dataset_v6.drop(dataset_v6[mask].index)
dataset_v7.shape

dataset_v7.loc[32,'to'] = "00" +  dataset_v7.loc[32,'to']  

mask = dataset_v7['to'].str.len() != 9
print(dataset_v7[mask].shape)
dataset_v7[mask]

dataset_v8 = dataset_v7.drop(dataset_v7[mask].index)
dataset_v8.shape

dataset_v8['from'] = dataset_v8['from'].apply(lambda t: t[2:])
dataset_v8['to'] = dataset_v8['to'].apply(lambda t: t[2:])


print(dataset_v8[dataset_v8['id'] == -1].shape)
print(dataset_v8[dataset_v8['id'] != -1].shape)

dataset_v8.to_csv('json_cleaned.csv', header=True, sep=',', index=False)
