==> 02_Lyon_EDA-FILLIN.ipynb <==
# Predicting the occupancies of Belgian trains

## 2. Exploratory Data Analysis


#vector/matrix library
import numpy as np
#data frame library (similar to R)
import pandas as pd

#visualization library
%matplotlib inline
import matplotlib.pyplot as plt
plt.style.use('ggplot')

#regular expression library for data cleasning
import re

dataset = pd.read_csv('json_cleaned.csv')
dataset.head(n=3)

### 2.1 Occupancy Distribution

* Make a pie chart showing the distrubution of the class labels (occupancy). Are they uniformly distributed, and if not what does that imply for future model performance?

HINT: use value_counts() and the pandas plot function for pie charts



dataset['occupancy']<FILL_IN>.plot(<FILL_IN>, figsize=(5,5))

### 2.2 Vehicle Distribution

* Same exercice but now for train types

dataset['train_type']<FILL_IN>.plot(<FILL_IN>)

## 2.3 Querytimes

1. Do query times between train and test set overlap? What does this imply?
2. Visualize the distribution of records over time (bar plots)
    * number of records per month
    * number of records per day => pay close attention, what does this chart teach you? 
    * number of records per weekday

HINT: pandas has plenty of support for time series operations. Have a look at the **resample** function, which allows you to aggregate over different time windows. It might be useful to override the current index with the query time values (a **datetime index** allows you to do time based slicing of the dataset, check set_index())

#1


#2a
dataset_ind = dataset.set_index(<FILL_IN>)

fig, ax = plt.subplots(1,1, figsize=(20,10))

#dataset_ind => use a datetime index
dataset_ind.resample(<FILL_IN>
ax.set_xticklabels(['Jul','Aug', 'Sep', 'Oct', 'Nov', 'Dec', 'Jan', 'Feb', 'Mar', 'Apr'])

fig, ax = plt.subplots(1,1, figsize=(20,10))

dataset_ind.resample(<FILL_IN>
ax.set_xticklabels([])

#2c
dataset['weekday'] = <FILL_IN>


fig, ax = plt.subplots(1,1, figsize=(20,10))

dataset.<FILL_IN>.plot

ax.set_xticklabels(['Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat', 'Sun'])

## 2.4 Querytimes (one more)

1. Visualize the distribution over hour of the day => what are you seeing
2. Plot this distribution for saturdays and sundays as well
3. What happens if you create the previous plot for [start, October] and [November, end] (and plot only weekdays) => what do you think that is happening? How would you fix it => fix it in the first notebook!!

#1
dataset['hour'] = <FILL_IN>

fig, ax = plt.subplots(1,1, figsize=(20,10))

dataset.<FILL_IN>.plot(kind='bar', ax=ax)

#2
fig, ax = plt.subplots(3,1, figsize=(20,20))

dataset<FILL_IN>.plot(kind='bar', ax=ax[0], title='weekday')
dataset<FILL_IN>.plot(kind='bar', ax=ax[1], title='saturday')
dataset<FILL_IN>.plot(kind='bar', ax=ax[2], title='sunday')

boundary = "20161031"

fig, ax = plt.subplots(2,1, figsize=(20,20))

dataset_weekdays = <FILL_IN>

dataset_weekdays<FILL_IN>.plot(kind='bar', ax=ax[0], title='weekdays [,Oct]')
dataset_weekdays<FILL_IN>.plot(kind='bar', ax=ax[1], title='weekdays [Oct,]')


## 2.5 Occupancy dependence

1. How does the occupancy depend on the train type? (use a normalized stacked bar chart)
2. How does the occupancy vary per month
3. How does the occupancy vary per weekday
4. How does the occupancy vary per hour (distinguish between weekday, saturday, sunday)

HINT: have a look at pivot_table functionality in pandas


#1
piv1 = dataset.pivot_table(<FILL_IN>

pivot_norm = <FILL_IN>

fig, ax = plt.subplots(1,1, figsize=(20,10))

pivot_norm.plot(kind='barh', stacked=True, ax=ax)

#2
piv2 = dataset.pivot_table(<FILL_IN>

fig, ax = plt.subplots(1,1, figsize=(20,10))

piv2.plot(kind='bar', stacked=True, ax=ax)

ax.set_xticklabels(['Jan', 'Feb', 'Jul','Aug', 'Sep', 'Oct', 'Nov', 'Dec'])

#3
piv3 = dataset.pivot_table(<FILL_IN>

fig, ax = plt.subplots(1,1, figsize=(20,10))

piv3.plot(kind='bar', stacked=True, ax=ax)

#4
fig, ax = plt.subplots(3,1, figsize=(20,20))

dataset.<FILL_IN>plot(kind='bar', stacked=True, ax=ax[0], title='weekday')
dataset.<FILL_IN>.fillna(0).plot(kind='bar', stacked=True, ax=ax[1], title='saturday')
dataset.<FILL_IN>.fillna(0).plot(kind='bar', stacked=True, ax=ax[2], title='sunday')

## If you would build a simple rule-based model (hand-crafted rules) how would it look like?

## I would use features for...

* FILL IN
* FILL IN
* FILL IN
* FILL IN
* FILL IN
* FILL IN
