==> 03_Lyon_ForecastingModel-FILL_IN.ipynb <==
# Predicting the occupancies of Belgian trains

#vector/matrix library
import numpy as np
#data frame library (similar to R)
import pandas as pd

#visualization library
%matplotlib inline
import matplotlib.pyplot as plt
plt.style.use('ggplot')

#regular expression library for data cleasning
import re

## 3. Building a basic forecasting model

### 3.1 Create a simple Kaggle submission

1. Create a csv document, just like the random.csv, submit to kaggle and try to get to the baseline performance in the leaderboard
2. **OPTIONAL:** what performance can you get by making a manually crafted rule-based prediction?

#1
dataset = pd.read_csv('json_cleaned.csv')

dataset['querytime'] = pd.to_datetime(dataset['querytime'])

mask = dataset['id'] < 0
training_set = dataset[mask]
test_set = pd.DataFrame(dataset[~mask])

submission = <FILL_IN>
submission.to_csv('submission1.csv', index=False)

### 3.2 Creating Numerical Features for sklearn

As mentioned in the theory slides: machine learning models can only deal with numerical data. In what follows we'll prepare a minimal dataframe and convert it to numpy arrays (which serve as input to an ML model)

Sklearn follows a very simple approach: http://scikit-learn.org/stable/data_transforms.html

- Estimators, have a fit method, which learns model parameters from a dataset
- Transformers have a transform method which applies this transformation model to unseen data. Making a prediction is considered a transformation, but also scaling the data for example!
- fit_transform combines the above
    
    
1. Prepare a dataframe with train_type, direction and a number of time series features (see EDA notebook). Make sure the data is scaled (for better model performance) and encode all categorical variables. Use the methods which are imported from the sklearn.preprocessing module

2. Split back into train and test and convert the dataframes to nd.array   



from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import LabelEncoder

#create additional time features: month, weekday, hour,... => some inspiration from the EDA notebook?
dataset_prep1 = <FILL_IN>

#encode train_direction
dataset_prep2 = <FILL_IN>

#map occupancies from [low, medium, high] -> [0,1,2]
dataset_prep3 = <FILL_IN>

#have a look at the pandas get dummies function, it converts categorical data into a numerical presentation...
dataset_prep4 = <FILL_IN>

#scale the time features for better model performance
dataset_prep5 = <FILL_IN>

#separate train from test
training_set = <FILL_IN>
test_set = <FILL_IN>

### 3.3 First Submission - Simple Linear Regression

* Fit a basic Linear Regresssion Model to the data
* Predict the labels for the test data
* The labels are continuous, choose a range to map to 0, 1 and 2 respectively
* Create a submission and upload to kaggle

from sklearn import linear_model

lin_reg = linear_model.LinearRegression()

X_train = <FILL_IN>
y_train = <FILL_IN>

#train model
lin_reg.<FILL_IN>

X_test = <FILL_IN>
y_pred = lin_reg.<FILL_IN>


def occ_to_int(o):
    <FILL_IN>

test_set['occupancy'] = test_set['occupancy'].apply(occ_to_int)

submission = <FILL_IN>
submission.to_csv('slr.csv', index=False)

### 3.4 Offline Model Evaluation

1. Make a (manual) Train / Test split 
    * First 90% of the training data is used for model training, the remainder is used for model evaluation
    * (K-fold cross validation is more common, do you have an idea why it might not be appropriate here?)
2. To evaluate the basic model, plot the <a href="http://scikit-learn.org/stable/modules/generated/sklearn.metrics.confusion_matrix.html"> confusion matrix </a> and caculate the accuracy
3. Test the previous model for different values of the hyperparameter
4. Try a number of different approaches, for example a <a href="http://scikit-learn.org/stable/modules/generated/sklearn.neighbors.KNeighborsClassifier.html"> k-Nearest Neighbours </a> classifier, definitely also have a look at a tree-based approach: <a href="http://scikit-learn.org/stable/modules/generated/sklearn.ensemble.RandomForestClassifier.html"> Random Forest Classifier </a>
5. Which of these models has the highest prediction accuracy? Are they complementary or do their correct predictions mostly overlap? (complementary models can be good candidates for building <a href="http://scikit-learn.org/stable/modules/ensemble.html">ensembles </a>)

6. Next steps for the quick students: 
    * use station locations and merge with to, from columns: stations.csv
    * use line stops: lineinfo.csv
    * look for holiday information

#1 split X_train in a part for offline training and a part for offline evaluation (=> you need labeled data!)
<FILL_IN>

#2
from sklearn.metrics import confusion_matrix

cfmatrix = <FILL_IN>

fig, ax = plt.subplots(1,1, figsize=(10,10))
cax = ax.imshow(confusion)
fig.colorbar(cax)
plt.xticks([0,1,2])
plt.yticks([0,1,2])

from sklearn.metrics import accuracy_score

accuracy = <FILL_IN>

#4
from sklearn.neighbors import KNeighborsClassifier
<FILL_IN>

rf.fit(X_tr, y_tr)
y_rf = rf.predict(X_te)
accuracy_score(y_te, y_rf)


#5 create datframe with all prediction and the actual solution
df_complementary = <FILL_IN>

print('NN correct: ' +str(<FILL_IN>))
print('RF correct: ' +str(<FILL_IN>))
print('LR correct: ' +str(<FILL_IN>))
print()
print('LR & NN correct: ' +str(<FILL_IN>))
print('LR & RF correct: ' +str(<FILL_IN>))
print('RF & NN correct: ' +str(<FILL_IN>))
